requirements.txt
Flask==2.2.2
Flask-Session==0.4.0
Flask-Login==0.6.2
Bootstrap-Flask==2.2.0
python-dateutil==2.8.2
mutagen==1.46.0
Pillow==9.4.0
